refactor:#33 분리된 nonMemberService 합침
